src/api/requirements.txt
azure-functions
pymssql>=2.2.8
openpyxl>=3.1.2
